+++ StudentSoup/src/main/java/ProjectDoge/StudentSoup/controller/boardreview/BoardReviewCallController.java
@@ -1,4 +1,4 @@
-package ProjectDoge.StudentSoup.controller.BoardReviewController;
+package ProjectDoge.StudentSoup.controller.boardreview;
 
 import ProjectDoge.StudentSoup.service.BoardReview.BoardReviewCallService;
 import lombok.RequiredArgsConstructor;
@@ -7,7 +7,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.concurrent.ConcurrentHashMap;
