+++ StudentSoup/src/main/java/ProjectDoge/StudentSoup/controller/boardreview/BoardReviewResisterController.java
@@ -1,4 +1,4 @@
-package ProjectDoge.StudentSoup.controller.BoardReviewController;
+package ProjectDoge.StudentSoup.controller.boardreview;
 
 import ProjectDoge.StudentSoup.dto.board.BoardReviewResDto;
 import ProjectDoge.StudentSoup.service.BoardReview.BoardReviewRegisterService;
